requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==6.0.2
pandas==2.3.3
pyarrow==22.0.0
numpy==2.3.4
python-dotenv==1.2.1